Move version.dart out of src
